Update Win32_NetworkAdapter
We already had a function that would return the active device IDs in an ArrayList. Now we have a function that uses that list to get the properties and their values as a HashMap<String,String>. I could've just printed the list but I intend to use this library in my GUI wrapper where a user would just select an ID and every property will appear in different text fields. The HashMap would help us achieve that.